+++ src/SFA.DAS.Payments.ScheduledJobs/Monitoring/ApprenticeshipData/CommitmentsDataContext.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace SFA.DAS.Payments.ScheduledJobs.ApprenticeshipsReferenceDataComparison
+namespace SFA.DAS.Payments.ScheduledJobs.Monitoring.ApprenticeshipData
 {
     public class CommitmentsDataContext : DbContext, ICommitmentsDataContext
     {
