+++ src/SFA.DAS.Payments.ScheduledJobs/Monitoring/ApprenticeshipData/IApprenticeshipsDataService.cs
@@ -1,0 +1,7 @@
+﻿namespace SFA.DAS.Payments.ScheduledJobs.Monitoring.ApprenticeshipData
+{
+    public interface IApprenticeshipsDataService
+    {
+        void ProcessComparison();
+    }
+}
