+++ src/SFA.DAS.Payments.ScheduledJobs/Monitoring/ApprenticeshipData/ICommitmentsDataContext.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace SFA.DAS.Payments.ScheduledJobs.ApprenticeshipsReferenceDataComparison
+namespace SFA.DAS.Payments.ScheduledJobs.Monitoring.ApprenticeshipData
 {
     public interface ICommitmentsDataContext
     {
